+++ tests/testthat/test-summary.R
@@ -1,9 +1,0 @@
-agg <- as_aggregation_structure(1:5)
-ind <- as_index(1:5)
-
-# These used to not work
-test_that("summary methods works", {
-  expect_no_condition(summary(agg))
-  expect_no_condition(summary(ind))
-  expect_no_condition(summary(elemental_index(1:5, contrib = TRUE)))
-})